% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidata-package.R
\docType{package}
\name{cidata}
\alias{cidata}
\alias{cidata-package}
\title{cidata: Data from Causal Inference by Hernán and Robins}
\description{
Includes 4 data sets used in Causal Inference by Hernán and Robins: 
  `nhefs`, a cleaned version of NHEFS data, `nhefs_complete`, a complete case 
  version of `nhefs`, `nhefs_codebook`, the codebook for NHEFS, and `greek_data`,
  the data from Table 2.2 used in Chapter 13.
}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolmbarrett@gmail.com}

}
\keyword{*}
\keyword{4}
\keyword{A}
\keyword{Causal}
\keyword{Data}
\keyword{Hernán}
\keyword{Inference}
\keyword{Package}
\keyword{Robins}
\keyword{and}
\keyword{by}
\keyword{cidata}
\keyword{cidata:}
\keyword{data}
\keyword{for}
\keyword{greek_data}
\keyword{includes}
\keyword{internal}
\keyword{nhefs}
\keyword{nhefs_codebook}
\keyword{nhefs_complete}
\keyword{sets:}
