% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhefs.R
\docType{data}
\name{nhefs_codebook}
\alias{nhefs_codebook}
\title{NHEFS Codebook}
\format{A data frame with 64 rows and 2 variables. Each row contains the name
of a variable (\code{variable}) in \code{\link{nhefs}} and \code{\link{nhefs_complete}} and a
description of what the variable represents (\code{description}).}
\source{
\url{wwwn.cdc.gov/nchs/nhanes/nhefs/}
}
\usage{
nhefs_codebook
}
\description{
\code{nhefs_codebook} is the codebook for \code{\link{nhefs}} and \code{\link{nhefs_complete}}.
}
\keyword{datasets}
